length_service/requirements.txt
fastapi
pydantic
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger